Address remaining review feedback on Process/ProcessStartInfo snippets
