+++ snippets/csharp/System.Diagnostics/Process/Overview/Project.csproj
@@ -3,4 +3,7 @@
     <OutputType>Exe</OutputType>
     <TargetFramework>net10.0</TargetFramework>
   </PropertyGroup>
+  <ItemGroup>
+    <Compile Remove="ArgsEcho/**/*.cs" />
+  </ItemGroup>
 </Project>
