+++ snippets/csharp/System.Diagnostics/Process/StandardOutput/Write500Lines/Project.csproj
@@ -3,7 +3,4 @@
     <OutputType>Exe</OutputType>
     <TargetFramework>net10.0</TargetFramework>
   </PropertyGroup>
-  <ItemGroup>
-    <Compile Remove="Write500Lines/**/*.cs" />
-  </ItemGroup>
 </Project>
