+++ snippets/csharp/System.Diagnostics/Process/StandardOutput/Write500Lines/write500lines.cs
@@ -1,17 +1,13 @@
-﻿using System;
+﻿// This example creates the Write500Lines.exe target used by the other examples in this section.
 
-public class Example3
+using System;
+
+for (int ctr = 0; ctr < 500; ctr++)
 {
-    public static void Run()
-    {
-        for (int ctr = 0; ctr < 500; ctr++)
-        {
-            Console.WriteLine($"Line {ctr + 1} of 500 written: {(ctr + 1) / 500.0:P2}");
-        }
-
-        Console.Error.WriteLine("\nSuccessfully wrote 500 lines.\n");
-    }
+    Console.WriteLine($"Line {ctr + 1} of 500 written: {(ctr + 1) / 500.0:P2}");
 }
+
+Console.Error.WriteLine("\nSuccessfully wrote 500 lines.\n");
 
 // The example displays the following output:
