+++ xml/System.Diagnostics/Process.xml
@@ -4568,7 +4568,7 @@ A thread in a process can migrate from processor to processor, with each migrati
 
 The last two examples in this section use the <xref:System.Diagnostics.Process.Start*> method to launch an executable named *Write500Lines.exe*. The following example contains its source code.
 
-:::code language="csharp" source="~/snippets/csharp/System.Diagnostics/Process/StandardOutput/write500lines.cs":::
+:::code language="csharp" source="~/snippets/csharp/System.Diagnostics/Process/StandardOutput/Write500Lines/write500lines.cs":::
 :::code language="fsharp" source="~/snippets/fsharp/System.Diagnostics/Process/StandardOutput/write500lines.fs":::
 :::code language="vb" source="~/snippets/visualbasic/api/system.diagnostics/process/standardoutput/write500lines.vb":::
 
@@ -4718,7 +4718,7 @@ You can use asynchronous read operations to avoid these dependencies and their d
 
 The last two examples in this section use the <xref:System.Diagnostics.Process.Start*> method to launch an executable named *Write500Lines.exe*. The following example contains its source code.
 
-:::code language="csharp" source="~/snippets/csharp/System.Diagnostics/Process/StandardOutput/write500lines.cs":::
+:::code language="csharp" source="~/snippets/csharp/System.Diagnostics/Process/StandardOutput/Write500Lines/write500lines.cs":::
 :::code language="fsharp" source="~/snippets/fsharp/System.Diagnostics/Process/StandardOutput/write500lines.fs":::
 :::code language="vb" source="~/snippets/visualbasic/api/system.diagnostics/process/standardoutput/write500lines.vb":::
