+++ xml/System.Diagnostics/ProcessStartInfo.xml
@@ -377,7 +377,7 @@ If you use this property to set command-line arguments, <xref:System.Diagnostics
 ## Examples
  The first example creates a small application (argsecho.exe) that echos its arguments to the console. The second example creates an application that invokes argsecho.exe to demonstrate different variations for the `Arguments` property.
 
- :::code language="csharp" source="~/snippets/csharp/System.Diagnostics/Process/Overview/processstartstatic3.cs":::
+ :::code language="csharp" source="~/snippets/csharp/System.Diagnostics/Process/Overview/ArgsEcho/processstartstatic3.cs":::
  :::code language="vb" source="~/snippets/visualbasic/System.Diagnostics/Process/Overview/processstartstatic3.vb":::
 
  :::code language="csharp" source="~/snippets/csharp/System.Diagnostics/Process/Overview/processstartstatic2.cs":::
@@ -1048,7 +1048,7 @@ The profile is unloaded after the new process and all child processes it has cre
 
 The last two examples in this section use the <xref:System.Diagnostics.Process.Start*> method to launch an executable named *Write500Lines.exe*. The following example contains its source code.
 
-:::code language="csharp" source="~/snippets/csharp/System.Diagnostics/Process/StandardOutput/write500lines.cs":::
+:::code language="csharp" source="~/snippets/csharp/System.Diagnostics/Process/StandardOutput/Write500Lines/write500lines.cs":::
 :::code language="vb" source="~/snippets/visualbasic/api/system.diagnostics/process/standardoutput/write500lines.vb":::
 
 The following example shows how to read from a redirected error stream and wait for the child process to exit. It avoids a deadlock condition by calling `p.StandardError.ReadToEnd` before `p.WaitForExit`. A deadlock condition can result if the parent process calls `p.WaitForExit` before `p.StandardError.ReadToEnd` and the child process writes enough text to fill the redirected stream. The parent process would wait indefinitely for the child process to exit. The child process would wait indefinitely for the parent to read from the full <xref:System.Diagnostics.Process.StandardError*> stream.
@@ -1189,7 +1189,7 @@ You can use asynchronous read operations to avoid these dependencies and their d
 
 The last two examples in this section use the <xref:System.Diagnostics.Process.Start*> method to launch an executable named *Write500Lines.exe*. The following example contains its source code.
 
-:::code language="csharp" source="~/snippets/csharp/System.Diagnostics/Process/StandardOutput/write500lines.cs":::
+:::code language="csharp" source="~/snippets/csharp/System.Diagnostics/Process/StandardOutput/Write500Lines/write500lines.cs":::
 :::code language="vb" source="~/snippets/visualbasic/api/system.diagnostics/process/standardoutput/write500lines.vb":::
 
 The following example shows how to read from a redirected stream and wait for the child process to exit. The example avoids a deadlock condition by calling `p.StandardOutput.ReadToEnd` before `p.WaitForExit`. A deadlock condition can result if the parent process calls `p.WaitForExit` before `p.StandardOutput.ReadToEnd` and the child process writes enough text to fill the redirected stream. The parent process would wait indefinitely for the child process to exit. The child process would wait indefinitely for the parent to read from the full <xref:System.Diagnostics.Process.StandardOutput*> stream.
